Move Nacho classes into respective packages
