+++ src/main/java/Nacho/Commands/HelpCommand.java
@@ -1,4 +1,8 @@
-public class HelpCommand implements Command{
+package Nacho.Commands;
+
+import Nacho.ChatContext;
+
+public class HelpCommand implements Command {
 
     @Override
     public void execute(String[] args, ChatContext context) {
