+++ src/main/java/Nacho/Commands/ListTasksCommand.java
@@ -1,4 +1,8 @@
-public class ListTasksCommand implements Command{
+package Nacho.Commands;
+
+import Nacho.ChatContext;
+
+public class ListTasksCommand implements Command {
     @Override
     public void execute(String[] args, ChatContext context) {
