+++ src/main/java/Nacho/Commands/UnmarkTaskCommand.java
@@ -1,4 +1,10 @@
-public class UnmarkTaskCommand implements Command{
+package Nacho.Commands;
+
+import Nacho.*;
+import Nacho.Exceptions.UserInputException;
+import Nacho.Tasks.Task;
+
+public class UnmarkTaskCommand implements Command {
 
     @Override
     public void execute(String[] args, ChatContext context) {
