+++ src/main/java/Nacho/Exceptions/UserInputException.java
@@ -1,6 +1,8 @@
+package Nacho.Exceptions;
+
 public class UserInputException extends RuntimeException{
     // Custom exception that gets thrown when user is missing mandatory fields in their commands
-    UserInputException(String message) {
+    public UserInputException(String message) {
         super(message);
     }
 }
