+++ src/main/java/Nacho/Tasks/TodoTask.java
@@ -1,0 +1,23 @@
+package Nacho.Tasks;
+
+public class TodoTask extends Task {
+
+    public TodoTask(String title) {
+        super(title);
+    }
+
+    @Override
+    public String getStorageRepresentation() {
+        String[] info = new String[3];
+        info[0] = "T";
+        info[1] = Integer.toString(this.isCompleted());
+        info[2] = this.getTitle();
+
+        return String.join(" | ", info);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
